requirements.txt
django==2.2
mysqlclient
pillow
qcloudsms-py
django-simple-captcha
django-pure-pagination
